requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# AWS & Storage
boto3==1.29.7
django-storages==1.14.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Media Processing
Pillow==10.0.1

# Environment Management
python-decouple==3.8

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development & Testing
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.5.2

# Email
django-anymail==10.2

# Forms & Widgets
django-widget-tweaks==1.5.0

# Utilities
django-extensions==3.2.3
django-cleanup==6.0.0
